pyproject.toml
[project]
name = "picassbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Raphael Jean", email = "raphael.jean.professionnel@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "tensorboard>=2.20.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
